+++ tracardi/process_engine/action/v1/connectors/hubspot/update_contact/plugin.py
@@ -55,6 +55,13 @@ class HubSpotContactUpdater(ActionRunner):
                     self.config.contact_id,
                     self.config.properties
                 )
+
+            if 'id' in result:
+                contact_id = result['id']
+                if not self.profile.metadata.system.has_integration('hubspot'):
+                    self.profile.metadata.system.set_integration('hubspot', contact_id)
+                    self.profile.set_updated_in_workflow()
+
             return Result(port="response", value=result)
 
         except HubSpotClientException as e: