+++ tracardi/process_engine/action/v1/connectors/mailchimp/add_to_audience/model/config.py
@@ -6,7 +6,7 @@ from tracardi.service.plugin.domain.config import PluginConfig
 
 class Config(PluginConfig):
     source: NamedEntity
-    list_id: str
+    list_id: NamedEntity
     email: str
     merge_fields: Dict[str, Any]
     subscribed: bool
@@ -15,7 +15,7 @@ class Config(PluginConfig):
     @field_validator("list_id")
     @classmethod
     def validate_list_id(cls, value):
-        if value is None or len(value) == 0:
+        if value is None or len(value.id) == 0:
             raise ValueError("This field cannot be empty.")
         return value