+++ tracardi/process_engine/action/v1/connectors/mailchimp/tag_contact/model/config.py
@@ -7,7 +7,7 @@ class Config(PluginConfig):
     source: NamedEntity
     email: str
     list_id: NamedEntity
-    delete: bool
+    tags: str
 
     @field_validator("email")
     @classmethod